pyproject.toml
[tool.poetry]
name = "django-app"
version = "0.1.0"
description = ""
authors = ["imbadr <tadjer.badr@gmail.com>"]
readme = "README.md"
packages = [{include = "django_app"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1.4"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"
psycopg2 = "^2.9.5"
djangorestframework-simplejwt = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"